pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "advanced-calculator"
version = "1.1.1"
description = "高级科学计算器，支持基本运算、科学函数、单位转换、复数运算等功能"
requires-python = ">=3.8"
dependencies = [
    "colorama>=0.4.6",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "packaging>=23.0",
]

[tool.pytest]
pythonpath = [
    ".",
    "src"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=html"

[project.scripts]
calc-cli = "src.calculator_cli:ScientificCalculator().run"
calc-gui = "src.gui_calculator:CalculatorGUI().run"